updates for NodeMCU 2.2.0 firmware

+++ src/variables_set.lua
@@ -1,14 +1,21 @@
-local me = { 
-  set = function (name, value)
-    local fn = name .. '.lua'
-    local f = file.open(fn, "w")
-    f.writeline("local " .. name .. " = " .. value)
-    f.writeline("return " .. name)
-    f.close()
-    node.compile(fn)
-    file.remove(fn)
-    print("Heap: ", node.heap(), "Wrote: ", fn)
-    collectgarbage() 
+local module = ...
+local function set(name, value)
+  if not value then return end
+  local fn = name .. '.lua'
+  local f = file.open(fn, "w")
+  f.writeline("local " .. name .. " = " .. value)
+  f.writeline("return " .. name)
+  f.close()
+  node.compile(fn)
+  file.remove(fn)
+  print("Heap: ", node.heap(), "Wrote: ", fn)
+  collectgarbage()
+end
+
+return function(name, value)
+  if package.loaded[module] then
+    package.loaded[module] = nil
   end
-}
-return me
+  module = nil
+  set(name, value)
+end
